+++ Client/IClient.cs
@@ -38,11 +38,11 @@ namespace Pathoschild.Http.Client
 
         /// <summary>Set whether HTTP error responses (e.g. HTTP 404) should be raised as exceptions by default.</summary>
         /// <param name="enabled">Whether to raise HTTP errors as exceptions by default.</param>
-        [Obsolete("Will be removed in version 4. Use `SetOptions` instead.")]
+        [Obsolete("Will be removed in 4.0. Use `" + nameof(SetOptions) + "` instead.")]
         IClient SetHttpErrorAsException(bool enabled);
 
         /// <summary>Set default options for all requests.</summary>
-        /// <param name="options">The options.</param>
+        /// <param name="options">The options to set. (Fields set to <c>null</c> won't change the current value.)</param>
         IClient SetOptions(FluentClientOptions options);
 
         /// <summary>Set the default user agent header.</summary>