+++ Tests/Client/RequestTests.cs
@@ -113,7 +113,7 @@ namespace Pathoschild.Http.Tests.Client
             // execute
             IRequest request = this
                 .ConstructRequest(methodName)
-                .WithOptions(new RequestOptions() { IgnoreNullArguments = ignoreNullArguments })
+                .WithOptions(new RequestOptions { IgnoreNullArguments = ignoreNullArguments })
                 .WithArgument(key, value);
 
             // verify
@@ -202,7 +202,7 @@ namespace Pathoschild.Http.Tests.Client
             // execute
             IRequest request = this
                 .ConstructRequest(methodName)
-                .WithOptions(new RequestOptions() { IgnoreNullArguments = ignoreNullArguments })
+                .WithOptions(new RequestOptions { IgnoreNullArguments = ignoreNullArguments })
                 .WithArguments(new[]
                 {
                     new KeyValuePair<string, object>(keyA, valueA),
@@ -227,7 +227,7 @@ namespace Pathoschild.Http.Tests.Client
             // execute
             IRequest request = this
                 .ConstructRequest(methodName)
-                .WithOptions(new RequestOptions() { IgnoreNullArguments = ignoreNullArguments })
+                .WithOptions(new RequestOptions { IgnoreNullArguments = ignoreNullArguments })
                 .WithArguments(new { keyA = valueA, keyB = valueB });
 
             // verify
@@ -445,7 +445,7 @@ namespace Pathoschild.Http.Tests.Client
             var client = new FluentClient("https://example.org", new HttpClient(mockHttp));
 
             // verify
-            IResponse response = await client.GetAsync("/").WithOptions(new RequestOptions() { IgnoreHttpErrors = true });
+            IResponse response = await client.GetAsync("/").WithOptions(new RequestOptions { IgnoreHttpErrors = true });
             Assert.NotNull(response, "The HTTP response is null.");
             Assert.NotNull(response.Message, "The HTTP response message is null.");
             Assert.AreEqual(HttpStatusCode.NotFound, response.Status, "The HTTP status doesn't match the response.");
@@ -680,6 +680,11 @@ namespace Pathoschild.Http.Tests.Client
             this.AssertEqual(request, new HttpMethod(method), uri, ignoreArguments);
         }
 
+        /// <summary>Assert that a query string is equal to the expected value (or excluded if null and <paramref name="ignoreNullArguments"/> is true).</summary>
+        /// <param name="arguments">The request arguments.</param>
+        /// <param name="key">The key to assert.</param>
+        /// <param name="value">The expected value.</param>
+        /// <param name="ignoreNullArguments">Whether null argument values should be ignored.</param>
         private void AssertQuerystringArgument(IDictionary<string, StringValues> arguments, string key, string value, bool ignoreNullArguments)
         {
             if (ignoreNullArguments && value == null)